fix(test): isolate shared path resolution
